backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0
pyjwt==2.9.0
cryptography==43.0.1
binance-connector==3.8.0
httpx==0.27.2
websockets==12.0
passlib[bcrypt]==1.7.4